git-svn-id: https://sources.ecopath.org/svn/Ecopath/trunk/Ecopath6@7622 a725721c-0ed6-864e-a6a8-02c2a1988c9d

+++ Sources/ScientificInterface/Ecosim/Tools/Shared/ucParmBlockCodes.vb
@@ -254,7 +254,7 @@ Namespace Ecosim
             If (Me.m_uic Is Nothing) Then Return
             If (Me.m_bInUpdate = True) Then Return
 
-            Me.SelectedBlock = CInt(Int(CSng(e.X) / Me.BlockWidth()))
+            Me.SelectedBlock = CInt(CSng(e.X) / Me.BlockWidth())
 
         End Sub
